requirements.txt
pycryptodome
tkinter
pycryptodome
streamlit